refactor(dashboard): Move dashboard into views folder
